+++ rust/grpc-api/src/lib.rs
@@ -1,1 +1,5 @@
 pub mod logging;
+pub mod services;
+
+pub use services::coordinator::CoordinatorService;
+pub use services::numprotoservice::NumProtoService;
