+++ rust/grpc-api/src/services/mod.rs
@@ -1,0 +1,2 @@
+pub mod coordinator;
+pub mod numprotoservice;
